AddCharacter update: id: AddCharacter.java --Overview-- This activity lets you add all of the string elements from the database the database, verifies that all of the fields aren't empty, and then sends them as a bundle to the AddSkills activity(not yet implemented). There they will be unpackaged and that activity will get the rest of the character information in order to add the character to the databse. If the user chooses wants to add an avatar to the character profile, then by pressing on the avatarPictureButton, they'll be able to change their avatar picture.
--functions--
-ImageButton: will allow you to upload a .jpeg file that will be used as the avatar picture. The button will send the .jpeg to the AddSkills
activity, and once a character has been successfully added, the .jpeg will be sent back to MainActivity where the ImageButton associated with
the profile being created will be updated with the file (FUNCTIONALITY NOT YET IMPLEMENTED)
id: avatarPictureButton

-8 TextViews: These are where the user will enter their information about their character. No error checking for characters witht the same
name in the database was done
ids: (from top to bottom) nameField, descriptionField, aspectsField, stuntsField, consequence1Field, consequence2Field, consequence3Field

- Back button: Used if the user chooses not to continue the character creation for this profile. Sends the user back to the main activity
with a boolean extra, false, that will make sure if the profile is selected on the main page again, it will take the user to the addCharacter
activity instead of the updateCharacter activity(NOT YET IMPLEMENTED)
id: backButton
OnClickListener: backListener

- Next button: This option makes sure all EditText fields aren't empty, if so the valid boolean will be marked false and whatever TextField
caused the error will display a toast about the error, and the associated TextField will be marked red until those changes are corrected.
If all of the fields are not empty, then the information from this activity will be sent as a bunlde (WITH THE CORRECT DATABASE NAMES) to the
AddSkills activity, where the rest of the database information will be retrieved.
id: nextButton
OnClickListener: nextListenerw